Bring the GDS VFD into the core library
Added the --with-cuda and --with-cufile flags to the autoconf build (both
must be given, but if CUDA is installed in /usr/local/cuda, they can be
the trivial form: "--with-cuda --with-cufile").  The GDS VFD will be
built if both are found.

And added HDF5_ENABLE_GDS_VFD:BOOL to the CMake build, which will use
the standard 'FindCUDAToolkit' package to locate the necessary libraries.

NOTE: the CMake build requires CMake 3.28 or greater in order to find
the cufile library correctly.

Signed-off-by: Quincey Koziol <qkoziol@nvidia.com>