h5cc shell portability fixes (#5465)
* h5cc: Fix bashism

Replace the `${@:2}` construct that is specific to bash shell with
a more portable approach based on `shift`, in order to restore
compatibility with more strict POSIX shells.  This fixes a regression
introduced in #5361.

* h5cc: Fix passing arguments to compiler subprocesses

Pass arguments to subprocesses via quoted `"$@"` rather than plain `$@`,
in order to prevent the shell from applying word splitting, filename
expansion, etc., and therefore ensure that they are passed through
to the compiler process unchanged.